+++ Block.h
@@ -46,7 +46,14 @@ public:
      */
     void incrementCount();
 
-    int getPartOfBlockContent(void* buffer, off_t offset, int count);
+    /**
+     *
+     * @param buffer
+     * @param offset
+     * @param count
+     * @return
+     */
+    int getPartOfBlockContent(void* buffer, off_t offset, size_t count);
 
 };
